image_analysis/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
python-dotenv==1.0.1
pydantic>=2.0.0
